backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
janome==0.5.0
pydantic==2.10.4
networkx==3.4.2
google-genai>=0.8.0
python-dotenv==1.0.1
